pyproject.toml
[tool.poetry]
name = "movie-review-project"
version = "0.1.0"
description = ""
authors = ["Haeebin <goqls4112@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
bs4 = "^0.0.2"
selenium = "^4.29.0"
apache-airflow = "^2.10.5"
pyspark = "^3.5.5"
pandas = "^2.2.3"
numpy = "^2.2.4"
google-cloud-storage = "^3.1.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.2"
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
